seperated the progress bar from main card component
